Add app files and move app and endpoints into app.js and import app into server.js
